this

+++ src/main/java/paul/fallen/clickgui/comp/CheckBox.java
@@ -21,7 +21,7 @@ public class CheckBox extends Comp {
     public void drawScreen(int mouseX, int mouseY) {
         super.drawScreen(mouseX, mouseY);
         //Gui.drawRect(parent.posX + x - 70, parent.posY + y, parent.posX + x + 10 - 70, parent.posY + y + 10,setting.getValBoolean() ? new Color(230,10,230).getRGB() : new Color(30,30,30).getRGB());
-        UIUtils.drawRect((int) (parent.posX + x - 70), (int) (parent.posY + y), 10, 10, setting.getValBoolean() ? new Color(230, 10, 230).getRGB() : new Color(30, 30, 30).getRGB());
+        UIUtils.drawRect(parent.posX + x - 70, parent.posY + y, 10, 10,setting.getValBoolean() ? new Color(230,10,230).getRGB() : new Color(30,30,30).getRGB());
         //Minecraft.getMinecraft().fontRendererObj.drawString(setting.getName(), (int)(parent.posX + x - 55), (int)(parent.posY + y + 1), new Color(200,200,200).getRGB());
         UIUtils.drawTextOnScreen(setting.getName(), (int)(parent.posX + x - 55), (int)(parent.posY + y + 1), new Color(200,200,200).getRGB());
     }
